Standardize SQL scripts and migrations: format index definitions, replace `REAL` data type with `NUMERIC`, adjust table and column references, and update associated tests and workflows.

+++ src/LogixDb.Data.Sqlite/Scripts/Migrations/Logix_202602131300_CreateTagProducerTable.sql
@@ -4,9 +4,9 @@ CREATE TABLE tag_producer
     produce_count      INTEGER NOT NULL,
     send_event_trigger INTEGER NOT NULL,
     unicast_permitted  INTEGER NOT NULL,
-    maximum_rpi        REAL    NOT NULL,
-    minimum_rpi        REAL    NOT NULL,
-    default_rpi        REAL    NOT NULL,
+    maximum_rpi        NUMERIC NOT NULL,
+    minimum_rpi        NUMERIC NOT NULL,
+    default_rpi        NUMERIC NOT NULL,
     record_hash        TEXT    NOT NULL,
     FOREIGN KEY (tag_id) REFERENCES tag (tag_id) ON DELETE CASCADE
 );
